add switch to activate "Throw Balls in Player" mode permanently and define the default ball size for this mode add an "Update All" button to the image manager
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2480 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ codeview.cpp
@@ -1680,7 +1680,7 @@ bool CodeViewer::ShowTooltip(SCNotification *pSCN)
 		if (t != string::npos)
 		{
 			const LRESULT linestart = SendMessage(m_hwndScintilla, SCI_POSITIONFROMLINE, CurrentLineNo, 0 );
-			if ( (wordstart - linestart) >= t ) return false;
+			if ( ((size_t)(wordstart - linestart)) >= t ) return false;
 		}
 	}
